misc mintests
