+++ tests/min_tests.py
@@ -30,7 +30,6 @@ def run_testsuit():
         "test_ahnet",
         "test_arraydataset",
         "test_auto3dseg",
-        "test_auto3dsegmentation",
         "test_cachedataset",
         "test_cachedataset_parallel",
         "test_cachedataset_persistent_workers",
